Issue #105
I attempted to fix this issue. However, I didn't find a sophisticated solution to the problem with the existing codes of the nav bar and footer (particularly the nav bar that changes its height). The posted solution will resolve the issue of seeing a white bar under the footer, but in some browsers and devices, there will be extra blanks on the top of the footer. 
Another possible solution I propose is to fix the footer on the window at the bottom, which means that the footer will no longer be at the end of the document but the bottom of the window. So people will always see the footer, and the contents can be scrolled down.